tests: split pkg/testutil/nettestutil
